pyproject.toml
[project]
name = "nc-spca"
version = "0.1.0"
description = "Gradient-Based Optimization for Network-Constrained Sparse PCA"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.2",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
    "pyrefly>=0.50.0",
    "ty>=0.0.14",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
nc-spca-run = "nc_spca.cli.run:main"
nc-spca-sweep = "nc_spca.cli.sweep:main"
nc-spca-reproduce = "nc_spca.cli.reproduce:main"
nc-spca-visualize = "nc_spca.cli.visualize:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
include = ["nc_spca*"]

[tool.pytest.ini_options]
pythonpath = [".", "src"]
